python/name-service/pyproject.toml
[tool.poetry]
name = "name-service"
version = "0.1.0"
description = ""
authors = ["Douglas Soo <dsoo@alum.mit.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
opentelemetry-api = "^1.8.0"
opentelemetry-sdk = "^1.8.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.8.0"
opentelemetry-instrumentation-flask = "^0.27b0"
opentelemetry-instrumentation-requests = "^0.27b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"